refactor: move gesture settings to new GestureSettingsPane
